Back Twice To Exit
_backTwice function added across all main screens
