+++ instashop/lib/ui/shop_page.dart
@@ -1,5 +1,7 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
-import 'package:instashop/widgets/customNavBar.dart';
+import 'package:instashop/widgets/custom_nav_bar.dart';
 import '../ui/item_in_category.dart';
 
 class ShopPage extends StatefulWidget {
@@ -10,10 +12,24 @@ class ShopPage extends StatefulWidget {
 }
 
 class _ShopPageState extends State<ShopPage> {
+  DateTime _lastQuitTime = DateTime(0);
+
+  Future<bool> _backTwice() async {
+    if (_lastQuitTime == null ||
+        DateTime.now().difference(_lastQuitTime).inSeconds > 1) {
+      print('Press again Back Button exit');
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text('Press back button again to exit')));
+      _lastQuitTime = DateTime.now();
+      return false;
+    } else {
+      print('sign out');
+      Navigator.of(context).pop(exit(0));
+    }
+  }
   @override
   Widget build(BuildContext context) {
     return WillPopScope(
-      onWillPop: () async => false,
+      onWillPop: _backTwice,
       child: new Scaffold(
           appBar: new AppBar(
             leading: new Container(),
